feat: real room loading — room.c/player.c/dread.c compiled, x64 file-image ports
The room-load state machine now runs the real decomp code end to end:
rm_*.rdx is decompressed (Expand), parsed (bhSetRoom), its models/motions
realized, textures decoded and uploaded, and player+weapon data loaded —
300s headless run reaches gameplay with no crash (127 texture slots live).

New real files in RECVX_GAME_SOURCES: room.c, objitm.c, eneset.c,
player.c, dread.c, ps2_NaColi.c (all zero PS2 asm).

x64 ports of PS2 32-bit file-image code:
- bhSetRoom: the 1004-byte RDX header is a PS2 ROM_WORK image (16 u32
  offsets + raw dmps + scalars). The PS2 in-place u32 relocation sweep
  shears every field on x64 — replaced with an explicit field-by-field
  conversion under RECVX_PC_PORT. The CUT_WORK array (32-bit cuttp,
  stride 0x2A8) gets a converted copy, allocated after the final
  memp=reladr rewind so later loads can't overwrite it.
- bhMnbBinRealize: serialized NJS_MOTION (12B) / NJS_MDATA2_MOD (2xu32)
  walked in place sheared fields and wrote 8-byte pointers over the next
  entry; now parses raw bytes and callocs a native NJS_MDATA2_MOD array
  (same strategy as the existing port bhMlbBinRealize).
- MN_WORK sizing: three hardcoded 12288 (= 512 * 24, the PS2 MN_WORK
  size) allocations in room.c/player.c now use sizeof(MN_WORK) * 512.
- Pointer-truncation aligns ((int)ptr) fixed in room.c, eneset.c and
  dread.c (uintptr_t, same pattern as system.c).
- dread.c: guarded op->mlwP derefs that PS2 tolerated as NULL-page reads
  (RX_MLWP_OK) — first weapon load runs before bhSetWeapon sets mlwP.

Support work:
- ninja_3d.c: real njSqrt/njInvertSqrt/njRotateXYZ (Z,Y,X order from
  ps2_NaMatrix.c:649)/njCalcVector/njUnitVector/njInnerProduct/
  njUnitRotPortion/njPushMatrixEx/njPopMatrixEx; removed bhKeepObjWork
  copy (real one now comes from dread.c).
- game_room_stubs.c (new, game target): no-op stubs for the enemy AI
  handlers (bhEne01..71), hitchk collision, lights, effects and weapon
  SE that the new files reference, plus real npCopyMemory/npSetMemoryL
  and a real njCheckPlane4IncludePoint (declared in ps2_NaColi.h but
  never defined in the decomp).
- Texture pool: 64 -> 512 slots (TEX_POOL_SLOTS + RX_GFX_TEX_SLOTS) — a
  real room exhausts 64 immediately; pool is not yet recycled per room.
- stub_game.c: removed stubs now provided by the real files (bhSetRDT,
  bhInitReadRDT, bhFinishRoom, bhPushGameData, bhInitPlayer,
  bhResetPlayer, bhStandPlayerMotion, bhCheckSubPack, bhInitObjItm,
  bhInitEnemy, bhReadPlayerData, bhReadWeaponData, WpnTab).

+++ src/ps2/veronica/prog/eneset.c
@@ -913,7 +913,7 @@ void* bhEne_CallocWork(int size, int param) // second parameter not present on D
         return NULL;
     }
     
-    sys->memp = (unsigned char*)(((unsigned int)sys->memp + 63) & ~0x3F);
+    sys->memp = (unsigned char*)(((uintptr_t)sys->memp + 63) & ~(uintptr_t)0x3F);
     
     ret_addr = sys->memp;
